Decimal to Fraction Converter
Handy pair of functions to help convert a decimal number into it's most reduced fractional form. The function decimalToFraction() takes in a decimal number and converts it into it's most immediate fractional form; the function fractionReducer() takes in the numerator and denominator supplied by decimalToFraction() and reduces it recursively as far as is possible.